CNDB-10221: use int for SAI postings (#1199)
After careful examination of the SAI code base, it is clear that we have a restriction that a single segment can have at most `Integer.MAX_VALUE` rows, and therefore, we can safely use `int` instead of `long` for postings. This PR proposes changes to make the code consistent, which should improve readability and future design, assuming that we want to continue to rely on having `Integer.MAX_VALUE` as the upper bound on number of rows in a segment.

The PR removes the `OffsetPostingList` because it incorrectly applies the segment offset in the `PostingList`, which should not happen.

Note that I was able to make all of the `long` to `int` conversions with only a handful of actual type conversions. I used `Math.toIntExact` to ensure we don't introduce any regressions. Additionally, I documented special cases that I thought were notable.

In the second commit, I updated part of the kd tree logic to account for the fact that doc ids (row ids) are always integers. We still use the `PackedLongValues` object though. Perhaps there is another data structure we could use?

+++ src/java/org/apache/cassandra/index/sai/disk/PostingList.java
@@ -26,26 +26,25 @@ import org.apache.cassandra.utils.Throwables;
 /**
  * Interface for advancing on and consuming a posting list.
  */
-//TODO Need to check int and long usage throughout this post DSP-19608
 @NotThreadSafe
 public interface PostingList extends Closeable
 {
     PostingList EMPTY = new EmptyPostingList();
 
-    long OFFSET_NOT_FOUND = -1;
-    long END_OF_STREAM = Long.MAX_VALUE;
+    int OFFSET_NOT_FOUND = -1;
+    int END_OF_STREAM = Integer.MAX_VALUE;
 
     @Override
     default void close() throws IOException {}
 
     /**
-     * Retrieves the next segment row ID, not including row IDs that have been returned by {@link #advance(long)}.
+     * Retrieves the next segment row ID, not including row IDs that have been returned by {@link #advance(int)}.
      *
      * @return next segment row ID
      */
-    long nextPosting() throws IOException;
+    int nextPosting() throws IOException;
 
-    long size();
+    int size();
 
     /**
      * @return {@code true} if this posting list contains no postings
@@ -67,7 +66,7 @@ public interface PostingList extends Closeable
      *
      * @return first segment row ID which is >= the target row ID or {@link PostingList#END_OF_STREAM} if one does not exist
      */
-    long advance(long targetRowID) throws IOException;
+    int advance(int targetRowID) throws IOException;
 
     /**
      * @return peekable wrapper of current posting list
@@ -82,14 +81,14 @@ public interface PostingList extends Closeable
         private final PostingList wrapped;
 
         private boolean peeked = false;
-        private long next;
+        private int next;
 
         public PeekablePostingList(PostingList wrapped)
         {
             this.wrapped = wrapped;
         }
 
-        public long peek()
+        public int peek()
         {
             if (peeked)
                 return next;
@@ -105,7 +104,7 @@ public interface PostingList extends Closeable
             }
         }
 
-        public long advanceWithoutConsuming(long targetRowID) throws IOException
+        public int advanceWithoutConsuming(int targetRowID) throws IOException
         {
             if (peek() == END_OF_STREAM)
                 return END_OF_STREAM;
@@ -119,7 +118,7 @@ public interface PostingList extends Closeable
         }
 
         @Override
-        public long nextPosting() throws IOException
+        public int nextPosting() throws IOException
         {
             if (peeked)
             {
@@ -130,13 +129,13 @@ public interface PostingList extends Closeable
         }
 
         @Override
-        public long size()
+        public int size()
         {
             return wrapped.size();
         }
 
         @Override
-        public long advance(long targetRowID) throws IOException
+        public int advance(int targetRowID) throws IOException
         {
             if (peeked && next >= targetRowID)
             {
@@ -158,19 +157,19 @@ public interface PostingList extends Closeable
     class EmptyPostingList implements PostingList
     {
         @Override
-        public long nextPosting() throws IOException
+        public int nextPosting() throws IOException
         {
             return END_OF_STREAM;
         }
 
         @Override
-        public long size()
+        public int size()
         {
             return 0;
         }
 
         @Override
-        public long advance(long targetRowID) throws IOException
+        public int advance(int targetRowID) throws IOException
         {
             return END_OF_STREAM;
         }
@@ -206,19 +205,19 @@ public interface PostingList extends Closeable
         }
 
         @Override
-        public long size()
+        public int size()
         {
             return delegate.size();
         }
 
         @Override
-        public long advance(long targetRowID) throws IOException
+        public int advance(int targetRowID) throws IOException
         {
             return delegate.advance(targetRowID);
         }
 
         @Override
-        public long nextPosting() throws IOException
+        public int nextPosting() throws IOException
         {
             return delegate.nextPosting();
         }
