+++ src/java/org/apache/cassandra/index/sai/disk/v1/postings/FilteringPostingList.java
@@ -57,11 +57,11 @@ public class FilteringPostingList implements PostingList
      * @return the segment row ID of the next match
      */
     @Override
-    public long nextPosting() throws IOException
+    public int nextPosting() throws IOException
     {
         while (true)
         {
-            long segmentRowId = delegate.nextPosting();
+            int segmentRowId = delegate.nextPosting();
 
             if (segmentRowId == PostingList.END_OF_STREAM)
             {
@@ -76,15 +76,15 @@ public class FilteringPostingList implements PostingList
     }
 
     @Override
-    public long size()
+    public int size()
     {
         return cardinality;
     }
 
     @Override
-    public  long advance(long targetRowID) throws IOException
+    public int advance(int targetRowID) throws IOException
     {
-        long segmentRowId = delegate.advance(targetRowID);
+        int segmentRowId = delegate.advance(targetRowID);
 
         if (segmentRowId == PostingList.END_OF_STREAM)
         {
@@ -92,7 +92,7 @@ public class FilteringPostingList implements PostingList
         }
 
         // these are always for leaf kdtree postings so the max is 1024
-        position = (int)delegate.getOrdinal();
+        position = delegate.getOrdinal();
 
         // If the ordinal of the ID we just read satisfies the filter, just return it...
         if (filter.get(position - 1))
